can now draw the textures onto images as well as a wirefram image.  bmp images are in BGR instead of RGB and need to be flipped
Added in normal vector calculation before putting texture on a md2
image.  Only works for BMP files atm, will beef this up soon!

+++ src/MD2Wrapper.h
@@ -25,8 +25,8 @@ struct VertexCoordinate
 
 struct TextureCoordinate
 {
-   float u;
-   float v;
+   short u;
+   short v;
 };
 
 struct Dimensions
